requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
gunicorn==23.0.0
Markdown==3.7
packaging==24.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.1
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
factory_boy==3.3.1
Faker==30.8.2
gunicorn==23.0.0
Markdown==3.7
packaging==24.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
